category proportional

+++ metamod/networks/linear_net.py
@@ -47,7 +47,7 @@ class LinearNet(BaseNetwork):
             y_pred = self.forward(x)
         else:
             y_pred = self.controlled_forward(x, g1_tilda=g1_tilda, g2_tilda=g2_tilda)
-        y_target = torch.from_numpy(y.T).to(self.device)
+        y_target = torch.from_numpy(y.T).type(self.dtype).to(self.device)
         loss1 = torch.sum((y_pred - y_target)**2)/(2.0*y.shape[0])
         loss2 = (self.reg_coef/2.0)*(torch.sum(self.W1**2) + torch.sum(self.W2**2))
         loss = loss1 + loss2
